+++ packages/expect/src/asymmetricMatchers.ts
@@ -179,6 +179,15 @@ class ObjectContaining extends AsymmetricMatcher<Record<string, unknown>> {
     } else {
       for (const property in this.sample) {
         if (
+          typeof this.sample[property] === 'object' &&
+          !(this.sample[property] instanceof AsymmetricMatcher)
+        ) {
+          this.sample[property] = objectContaining(
+            this.sample[property] as Record<string, unknown>,
+          );
+        }
+
+        if (
           !hasProperty(other, property) ||
           !equals(this.sample[property], other[property])
         ) {