+++ packages/jest-validate/src/validate.ts
@@ -14,8 +14,8 @@ let hasDeprecationWarnings = false;
 const shouldSkipValidationForPath = (
   path: Array<string>,
   key: string,
-  blacklist?: Array<string>,
-) => (blacklist ? blacklist.includes([...path, key].join('.')) : false);
+  denylist?: Array<string>,
+) => (denylist ? denylist.includes([...path, key].join('.')) : false);
 
 const _validate = (
   config: Record<string, any>,
@@ -70,7 +70,11 @@ const _validate = (
         options.error(key, config[key], exampleConfig[key], options, path);
       }
     } else if (
-      shouldSkipValidationForPath(path, key, options.recursiveBlacklist)
+      shouldSkipValidationForPath(
+        path,
+        key,
+        options.recursiveDenylist || options.recursiveBlacklist,
+      )
     ) {
       // skip validating unknown options inside blacklisted paths
     } else {
@@ -81,8 +85,12 @@ const _validate = (
     if (
       options.recursive &&
       !Array.isArray(exampleConfig[key]) &&
-      options.recursiveBlacklist &&
-      !shouldSkipValidationForPath(path, key, options.recursiveBlacklist)
+      (options.recursiveDenylist || options.recursiveBlacklist) &&
+      !shouldSkipValidationForPath(
+        path,
+        key,
+        options.recursiveDenylist || options.recursiveBlacklist,
+      )
     ) {
       _validate(config[key], exampleConfig[key], options, [...path, key]);
     }
@@ -101,16 +109,16 @@ const validate = (
 ): {hasDeprecationWarnings: boolean; isValid: boolean} => {
   hasDeprecationWarnings = false;
 
-  // Preserve default blacklist entries even with user-supplied blacklist
-  const combinedBlacklist: Array<string> = [
-    ...(defaultConfig.recursiveBlacklist || []),
-    ...(options.recursiveBlacklist || []),
+  // Preserve default denylist entries even with user-supplied denylist
+  const combinedDenylist: Array<string> = [
+    ...(defaultConfig.recursiveDenylist || []),
+    ...(options.recursiveDenylist || options.recursiveBlacklist || []),
   ];
 
   const defaultedOptions: ValidationOptions = Object.assign({
     ...defaultConfig,
     ...options,
-    recursiveBlacklist: combinedBlacklist,
+    recursiveDenylist: combinedDenylist,
     title: options.title || defaultConfig.title,
   });