+++ packages/jest-worker/src/workers/__tests__/NodeThreadsWorker.test.js
@@ -150,11 +150,11 @@ it('provides stdout and stderr from the threads', async () => {
   const stdout = worker.getStdout();
   const stderr = worker.getStderr();
 
-  worker._worker.stdout.end('Hello ', {encoding: 'utf8'});
-  worker._worker.stderr.end('Jest ', {encoding: 'utf8'});
+  worker._worker.stdout.end('Hello ', 'utf8');
+  worker._worker.stderr.end('Jest ', 'utf8');
   worker._worker.emit('exit');
-  worker._worker.stdout.end('World!', {encoding: 'utf8'});
-  worker._worker.stderr.end('Workers!', {encoding: 'utf8'});
+  worker._worker.stdout.end('World!', 'utf8');
+  worker._worker.stderr.end('Workers!', 'utf8');
   worker._worker.emit('exit', 0);
 
   await expect(getStream(stdout)).resolves.toEqual('Hello World!');