+++ e2e/__tests__/v8Coverage.test.ts
@@ -6,24 +6,33 @@
  */
 
 import * as path from 'path';
-import {onNodeVersions} from '@jest/test-utils';
 import runJest from '../runJest';
+import {runYarn} from '../Utils';
 
 const DIR = path.resolve(__dirname, '../v8-coverage');
 
-onNodeVersions('>=10', () => {
-  test('prints coverage', () => {
-    const sourcemapDir = path.join(DIR, 'no-sourcemap');
+test('prints coverage', () => {
+  const sourcemapDir = path.join(DIR, 'no-sourcemap');
 
-    const {stdout, exitCode} = runJest(
-      sourcemapDir,
-      ['--coverage', '--coverage-provider', 'v8'],
-      {
-        stripAnsi: true,
-      },
-    );
+  const {stdout, exitCode} = runJest(
+    sourcemapDir,
+    ['--coverage', '--coverage-provider', 'v8'],
+    {stripAnsi: true},
+  );
 
-    expect(exitCode).toBe(0);
-    expect(stdout).toMatchSnapshot();
-  });
+  expect(exitCode).toBe(0);
+  expect(stdout).toMatchSnapshot();
+});
+
+test('vm script coverage generater', () => {
+  const dir = path.resolve(__dirname, '../vmscript-coverage');
+  runYarn(dir);
+  const {stdout, exitCode} = runJest(
+    dir,
+    ['--coverage', '--coverage-provider', 'v8'],
+    {stripAnsi: true},
+  );
+
+  expect(exitCode).toBe(0);
+  expect(stdout).toMatchSnapshot();
 });