+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - `[jest-config]` Simplify transform RegExp ([#10207](https://github.com/facebook/jest/pull/10207))
 - `[jest-fake-timers]` Lazily instantiate mock timers ([#10551](https://github.com/facebook/jest/pull/10551))
 - `[jest-runtime]` `require.main` is no longer `undefined` when using `jest.resetModules` ([#10626](https://github.com/facebook/jest/pull/10626))
+- `[jest-runtime]` Keep the original appearance in the vm.script code-coverage. ([#10645](https://github.com/facebook/jest/pull/10645))
 - `[@jest/types]` Add missing values for `timers` ([#10632](https://github.com/facebook/jest/pull/10632))
 
 ### Chore & Maintenance